+++ src/editor/actions/sagas/__tests__/eval-queue-saga.test.js
@@ -13,7 +13,7 @@ import {
   loadingLanguageConsoleMsg,
   addInputToConsole,
   evalTypeConsoleError
-} from "../console-message-actions";
+} from "../../console-message-actions";
 
 // this string is the returnValue from a saga was still running when
 // it was timed out by redux-test-plan
