+++ src/editor/components/menu/__tests__/notebook-task-button.test.js
@@ -4,7 +4,7 @@ import { shallow } from "enzyme";
 import IconButton from "@material-ui/core/IconButton";
 
 import NotebookTaskButton from "../notebook-task-button";
-import UserTask from "../../../actions/user-task";
+import UserTask from "../../../user-tasks/user-task";
 
 describe("NotebookTaskButton has one IconButton", () => {
   const nbTask = shallow(
