+++ src/editor/components/menu/editor-mode-controls.jsx
@@ -6,7 +6,7 @@ import FastForward from "@material-ui/icons/FastForward";
 import EditorToolbarMenu from "./editor-toolbar-menu";
 import NotebookTaskButton from "./notebook-task-button";
 
-import tasks from "../../actions/task-definitions";
+import tasks from "../../user-tasks/task-definitions";
 
 export default class EditorModeControls extends React.Component {
   render() {
