+++ src/editor/user-tasks/external-link-task.js
