+++ src/eval-frame/actions/eval-frame-tasks.js
@@ -1,4 +1,4 @@
-import UserTask from "../../editor/actions/user-task";
+import UserTask from "../../editor/user-tasks/user-task";
 import { postKeypressToEditor } from "../port-to-editor";
 
 const tasks = {};
