+++ src/eval-frame/components/panes/console/history-input-item.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import styled from "react-emotion";
-import DoubleChevronIcon from "../double-chevron-icon";
+import DoubleChevronIcon from "./double-chevron-icon";
 import BaseIcon from "./base-icon";
 import ConsoleContainer from "./console-container";
 import ConsoleGutter from "./console-gutter";
